+++ headers/quien_empieza.h
@@ -1,0 +1,10 @@
+#pragma once
+#ifndef QUIEN_EMPIEZA_H
+#define QUIEN_EMPIEZA_H
+
+#include "../quien_empieza.cpp"
+
+void quien_empieza(Jugadores *jugador, int &jugadorActual);
+void obtener_quien_empieza(Jugadores *jugador, int &jugadorActual, int jugadorDadoMaximo, int jugadorSumaMaxima);
+
+#endif //QUIEN_EMPIEZA_H
